Pass 'true' to base class AIStatefulTask.
When you derive A from a class 'B' and want/need to
pass parameters to the constructor of B, you just
include it in the initializer list of A:

    class A : public B {
      public:
        A() : B(true), m_other(3), ...

In this case, AIStatefulTask takes a boolean, but only
when (the macro) CWDEBUG is defined.

In order to avoid the ugly:

    class Bunny : public AIStatefulTask {
      public:
        Bunny() :
    #ifdef CWDEBUG
            B(true),
    #endif
            m_radius(0.02), ...

'cwds/debug.h' defines a macro called DEBUG_ONLY() that
we can use. Either:

        Bunny() : B(DEBUG_ONLY(true)), m_other(0.02), ...

or

        Bunny() : DEBUG_ONLY(B(true),) m_radius(0.02), ...

would work, but the latter is clearly more ugly because
it includes the comma inside the macro.